Fixed ACCEL_AXIS_* defines
The previous value were not working because 'LIS3DH_writeAxisDataForAxisSerial' and 'LIS3DH_writeAxisDataForAxisWifi' truncate values > 2, ie:
if (axis > 2) axis = 0;

Therefore, the value sent was always the 'X' channel. Moreover, the previous values also have an other side effect. Indeed, 'sampleCounter' is coded on a byte, and therefore it's value goes from 0 to 127. With the previous values of ACCEL_AXIS_*, add the end of an iteration for ids, only the X channel was sent because 'sampleCounter % 10' never reached ACCEL_AXIS_Y and ACCEL_AXIS_Z because values 128 and 129 are not possible for ids.